Testing done for now; database appears to be working as expected.
	- Refactoring and slight cleanup done on database
	- Additional tests written and performed
	- De/serialization of Time objects changed to use rational numbers because of differing precisions and comparisons.
	- Testing class Foo implemented with `#hash` to account for comparing hashes in testing.
	- and many other things.

Next up: cleanup of database code, particularly with regards to the ranges being used in selecting arrays.
