+++ table_foosball/database_test.rb
@@ -10,9 +10,17 @@ class DatabaseTest < Minitest::Test
     end
 
     def ==(other)
-      self.instance_variables.each do |variable_name|
-        self.instance_variable_get(variable_name) == other.instance_variable_get(variable_name)
+      instance_variables.each do |variable_name|
+        return false unless instance_variable_get(variable_name) == other.instance_variable_get(variable_name)
       end
+
+      true
+    end
+
+    alias eql? ==
+
+    def hash
+      instance_variables.map { |name| [name, instance_variable_get(name)] }.hash
     end
   end
 
@@ -40,38 +48,92 @@ class DatabaseTest < Minitest::Test
   end
 
   def teardown
-    File.open("foo_out.txt", "w") { |f| f.write(File.read("foo_test.sdb")) }
     File.delete("./foo_test.sdb", "./bar_test.sdb")
   end
 
   def test_create_new_database
-    name = :test_database
-    test_database = Database.new(name)
+    name = :test_db
+    test_db = Database.new(name)
 
-    assert_equal Database, test_database.class
+    assert_equal Database, test_db.class
   end
 
-  def test_save_and_load_foos_on_database
-    assert_equal "", File.read("./foo_test.sdb")
-    assert_nil @foodb.load
+  def test_save_load_string
+    @foodb.save("sebastian")
 
+    assert_equal "sebastian", @foodb.load
+  end
+
+  def test_save_load_time
+    
+  end
+
+  def test_save_load_array
+    arr = ["sebastian"]
+
+    @foodb.save(arr)
+
+    assert_equal arr, @foodb.load
+  end
+
+  def test_save_load_hash
+    hash = {}
+    
+    hash[:key1] = "foo"
+    hash[:key2] = "bar"
+    hash[:key3] = "foobar"
+    hash[:key4] = { :foz => :baz }
+    hash[:key6] = ["hey", "you", "there"]
+    hash[:key7] = { ["element1", :element2, { @bar1 => :element3 } ] => { Time.now => { @foo1 => 2 } } }
+
+    @foodb.save(hash)
+
+    assert_equal hash, @foodb.load
+  end
+
+  def test_save_object_with_instance_variables
+    @bardb.save(@bar1)
+
+    assert_equal @bar1, @bardb.load
+  end
+
+  def test_save_array_of_simple_objects
+    objs = ["hey", "ho", ["let's", "go"], 2]
+    
+    @foodb.save(objs)
+    
+    assert_equal objs, @foodb.load
+  end
+
+  def test_save_array_of_objects
+    bars = [@bar1, @bar2]
+
+    @bardb.save(bars)
+
+    assert_equal bars, @bardb.load
+  end
+
+  def test_save_foo_with_bars_as_instance_variables
     @foodb.save(@foos)
 
-    refute_equal "", File.read("./foo_test.sdb")
-    
-    @foos.zip(@foodb.load).each do |exp_foo, act_foo|
-      assert_true exp_foo == act_foo
-    end
+    assert_equal @foos, @foodb.load
   end
 
-  def test_save_and_load_bars_on_database
-    assert_equal "", File.read("./bar_test.sdb")
-    assert_nil @bardb.load
+  def test_save_load_nil
+    @foodb.save(nil)
 
-    @bardb.save(@bars)
+    assert_equal nil, @foodb.load
+  end
 
-    refute_equal "", File.read("./bar_test.sdb")
-    
-    assert @bars == @bardb.load
+  def test_save_empty_array
+    @foodb.save([])
+
+    assert_equal [], @foodb.load
+  end
+
+  def test_save_empty_hash
+    @foodb.save({})
+
+    assert_equal({}, @foodb.load)
   end
 end
